Added Service for module for getting of configuration of provider and removed this functionality from Controller.
